requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.9.2
PyYAML==6.0.2
httpx==0.27.2
redis==5.0.8
APScheduler==3.10.4
pytrends==4.9.2
beautifulsoup4==4.12.3
python-dotenv==1.0.1
loguru==0.7.2
sendgrid==6.11.0
twilio==9.2.3
slack-sdk==3.33.1
